+++ Remote/Data/Settings.cs
@@ -17,5 +17,9 @@
     public int SelectedTabIndex { get; set; } = 0;
     public bool ShaderFileChecked { get; set; } = true;
     public bool WrapChecked { get; set; } = true;
+
+    public decimal VisIntensity { get; set; } = 1.0M;
+    public decimal VisShift { get; set; } = 0.0M;
+    public int VisVersion { get; set; } = 1;
   }
 }